nls: cast the _() macro argument like gettext() already does
_(a) forwarded its argument straight to etos_gettext(unsigned long)
without a cast, an implicit pointer-to-integer conversion at every
_("literal") call site once CONF_WITH_NLS is on. None of the current
CI configs build with CONF_WITH_NLS, so this went unnoticed by CI.

+++ include/nls.h
@@ -21,7 +21,7 @@
 
 #define NLS_MSGID_FLAG 0x80000000UL
 
-# define _(a) etos_gettext(a)
+# define _(a) etos_gettext((unsigned long)(a))
 # define N_(a) ((char *)((unsigned long)(a) | NLS_MSGID_FLAG))
 # define gettext(a) etos_gettext((unsigned long)(a))
 const char *etos_gettext(unsigned long msgid);
